+++ src/main/java/net/rentalhost/plugins/php/hammer/inspections/codeStyle/ParameterDefaultsNullInspection.kt
@@ -71,7 +71,6 @@ class ParameterDefaultsNullInspection: PhpInspection() {
 
             enforcesNullableType(project)
 
-
             createAssignment(project, parameterDefaultValue)
         }