+++ src/test/java/net/rentalhost/plugins/php/hammer/inspections/deadCode/UnusedUseVariableInspectionTestCase.kt
@@ -1,0 +1,7 @@
+package net.rentalhost.plugins.php.hammer.inspections.deadCode
+
+import net.rentalhost.plugins.services.TestCase
+
+class UnusedUseVariableInspectionTestCase: TestCase() {
+    fun testAll(): Unit = testInspection(UnusedUseVariableInspection::class.java)
+}